pyproject.toml
[project]
name = "auto-tasker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "google-api-python-client>=2.183.0",
    "google-auth-httplib2>=0.2.0",
    "google-auth-oauthlib>=1.2.2",
    "gspread>=6.2.1",
    "litellm>=1.77.5",
    "matplotlib>=3.10.7",
    "openai-agents>=0.3.2",
    "pandas>=2.3.3",
    "reportlab>=4.4.4",
]
